+++ models/staging/sap_adventure_works/stg_sap__location.sql
@@ -3,8 +3,8 @@ with
         select
             cast(locationid as int) as locationid
             , cast(name as string) as name
-            , cast(costrate as float) as costrate
-            , cast(availability as float) as availability
+            , cast(costrate as decimal) as costrate
+            , cast(availability as decimal) as availability
             , cast(modifieddate as timestamp) as modifieddate
         from {{ source('sap_adventure_works', 'location') }}
     )
