+++ models/staging/sap_adventure_works/stg_sap__salestaxrate.sql
@@ -4,7 +4,7 @@ with
             cast(salestaxrateid as int) as salestaxrateid
             , cast(stateprovinceid as int) as stateprovinceid
             , cast(taxtype as int) as taxtype
-            , cast(taxrate as float) as taxrate
+            , cast(taxrate as decimal) as taxrate
             , cast(name as string) as name
             , cast(rowguid as string) as rowguid
             , cast(modifieddate as timestamp) as modifieddate
